move sign by secondary key to flaky tests
