+++ test/source/tests/tests/compose.ts
@@ -420,22 +420,6 @@ export const defineComposeTests = (testVariant: TestVariant, testWithBrowser: Te
       await composePage.notPresent('@action-auth-with-contacts-scope');
     }));
 
-    ava.default('[compose[global:compatibility] - standalone - different send from, new signed message, verification in mock', testWithBrowser('compatibility', async (t, browser) => {
-      const settingsPage = await browser.newPage(t, '/chrome/settings/modules/add_key.htm?acctEmail=flowcrypt.compatibility%40gmail.com&parent_tab_id=0');
-      const key = Config.key('flowcryptcompatibility.from.address');
-      await settingsPage.waitAndClick('#source_paste');
-      await settingsPage.waitAndType('.input_private_key', key.armored!);
-      await settingsPage.waitAndClick('#toggle_input_passphrase');
-      await settingsPage.waitAndType('#input_passphrase', key.passphrase!);
-      await settingsPage.waitAndClick('.action_add_private_key', { delay: 1 });
-      await settingsPage.waitTillGone('.featherlight.featherlight-iframe'); // dialog closed
-      await settingsPage.close();
-      const composePage = await ComposePageRecipe.openStandalone(t, browser, 'compatibility');
-      await composePage.selectOption('@input-from', 'flowcryptcompatibility@gmail.com');
-      await ComposePageRecipe.fillMsg(composePage, { to: 'human@flowcrypt.com' }, 'New Signed Message (Mock Test)', { encrypt: false });
-      await ComposePageRecipe.sendAndClose(composePage);
-    }));
-
     ava.default('[compose[global:compatibility]] - standalone - new message, open footer', testWithBrowser('compatibility', async (t, browser) => {
       const composePage = await ComposePageRecipe.openStandalone(t, browser, 'compatibility');
       await ComposePageRecipe.fillRecipients(composePage, { to: 'human@flowcrypt.com' }, 'new');
